Move dummy data to temp folder
