+++ src/components/kanban/KanbanBoard.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+import { defaultCols, defaultTasks } from "@/temp/constants";
+import { Column, Id, Task } from "@/temp/types";
 import {
   DndContext,
   DragEndEvent,
@@ -17,103 +19,6 @@ import KanbanCard from "./KanbanCard";
 import KanbanColumn from "./KanbanColumn";
 
 interface KanbanBoardProps {}
-
-export type Id = string | number;
-
-export type Column = {
-  id: Id;
-  title: string;
-};
-
-export type Task = {
-  id: Id;
-  columnId: Id;
-  content: string;
-};
-
-const defaultCols: Column[] = [
-  {
-    id: "todo",
-    title: "Todo",
-  },
-  {
-    id: "doing",
-    title: "Work in progress",
-  },
-  {
-    id: "done",
-    title: "Done",
-  },
-];
-
-const defaultTasks: Task[] = [
-  {
-    id: "1",
-    columnId: "todo",
-    content: "List admin APIs for dashboard",
-  },
-  {
-    id: "2",
-    columnId: "todo",
-    content:
-      "Develop user registration functionality with OTP delivered on SMS after email confirmation and phone number confirmation",
-  },
-  {
-    id: "3",
-    columnId: "doing",
-    content: "Conduct security testing",
-  },
-  {
-    id: "4",
-    columnId: "doing",
-    content: "Analyze competitors",
-  },
-  {
-    id: "5",
-    columnId: "done",
-    content: "Create UI kit documentation",
-  },
-  {
-    id: "6",
-    columnId: "done",
-    content: "Dev meeting",
-  },
-  {
-    id: "7",
-    columnId: "done",
-    content: "Deliver dashboard prototype",
-  },
-  {
-    id: "8",
-    columnId: "todo",
-    content: "Optimize application performance",
-  },
-  {
-    id: "9",
-    columnId: "todo",
-    content: "Implement data validation",
-  },
-  {
-    id: "10",
-    columnId: "todo",
-    content: "Design database schema",
-  },
-  {
-    id: "11",
-    columnId: "todo",
-    content: "Integrate SSL web certificates into workflow",
-  },
-  {
-    id: "12",
-    columnId: "doing",
-    content: "Implement error logging and monitoring",
-  },
-  {
-    id: "13",
-    columnId: "doing",
-    content: "Design and implement responsive UI",
-  },
-];
 
 const KanbanBoard = ({}: KanbanBoardProps) => {
   const [tasks, setTasks] = useState<Task[]>(defaultTasks);
