+++ src/temp/types.ts
@@ -1,0 +1,12 @@
+export type Id = string | number;
+
+export type Column = {
+  id: Id;
+  title: string;
+};
+
+export type Task = {
+  id: Id;
+  columnId: Id;
+  content: string;
+};
